[API-97] Split api/data/weather/index.ts into fetch/parse/aggregate; index.ts orchestrates.
